backend/requirements.txt
# ── Navisense Backend Dependencies ────────────────────────────────────────────
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data Validation
pydantic==2.10.3

# HTTP & Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Gemini AI
google-generativeai==0.8.3

# Environment Variables
python-dotenv==1.0.1
